refactor into pokemon module
