+++ README.md
@@ -1,2 +1,17 @@
 # rust-pokemon-cli
 A CLI tool using the PokeApi used to generate tables and stats blocks for TTRPGs
+
+
+## Use Case
+
+
+/* Example Output
+
+| NAME      | TYPES        | ABILITIES    | FLAVOR TEXT    |
+| --------- | ------------ | ------------ | -----------    |
+| Bulbasaur | grass,poison | growl,tackle | blah blah blah |
+| Rhyhorn   | ground,rock  | fart,giggle  | blah blah blah |
+
+
+
+*/
